[#23] Refactor: 리뷰 반영으로 인한 리팩토링
